change measurements script

+++ measurements.sh
@@ -1,16 +1,28 @@
 # echo "Running sequential\n"
 # ./HuffmanProject samples/1000000000.txt binary seq
+file_name=$1
+# check if file exists
+if [ ! -f "$file_name" ]
+then
+    echo "File $file not found!"
+    exit 1
+fi
+
+rm measurements/seq.csv
+echo "Running sequential\n"
+./HuffmanProject $file_name binary seq
+
 rm measurements/threads.csv
 for i in 2 4 8 16 32 64 128
 do
     echo "Running with $i threads\n"
-    ./HuffmanProject samples/1000000000.txt binary t $i
+    ./HuffmanProject $file_name binary t $i
 done
 
 rm measurements/ff.csv
 for i in 2 4 8 16 32 64 128
 do
     echo "Running with $i threads\n"
-    ./HuffmanProject samples/1000000000.txt binary ff $i
+    ./HuffmanProject $file_name binary ff $i
 done
